+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.0
 
 - Add golden tests
+- Add optional ScaffoldBuilder that can be used to provide a custom app scaffold around all pages
 
 ## 1.1.2